rename hierarchy in worker.allocation
