+++ core/src/main/java/tachyon/worker/allocation/AllocateRR.java
@@ -15,7 +15,7 @@
 
 package tachyon.worker.allocation;
 
-import tachyon.worker.hierarchy.StorageDir;
+import tachyon.worker.tiered.StorageDir;
 
 /**
  * Allocate space on a StorageDir selected in round robin from StorageDirs.
