+++ core/src/main/java/tachyon/worker/allocation/AllocateRandom.java
@@ -17,7 +17,7 @@ package tachyon.worker.allocation;
 
 import java.util.Random;
 
-import tachyon.worker.hierarchy.StorageDir;
+import tachyon.worker.tiered.StorageDir;
 
 /**
  * Allocate space on a StorageDir that has enough space and is selected randomly.
